Initial funcitonality on esp32
Use shared epd code to handle the display.
Use PNG component for decoding, no need to use miniz.c because I think
the linker is able to find the one already in ROM and use that.

Now able to download and display appropriate PNG files!

After wich, sit in a loop and count ticks. We should really set a future alarm
and deepsleep, but this a good start.